requirements.txt
flask
opencv-python
numpy
scikit-learn
joblib
torch
torchvision
ultralytics<=8.3.40
supervision 
roboflow
python-dotenv
firebase-admin
flask_cors
twilio